+++ apps/web/app/(use-page-wrapper)/availability/[schedule]/page.tsx
@@ -1,3 +1,5 @@
 "use client";
 
-export { AvailabilitySettingsWebWrapper } from "~/availability/[schedule]/schedule-view";
+import { AvailabilitySettingsWebWrapper } from "~/availability/[schedule]/schedule-view";
+
+export default AvailabilitySettingsWebWrapper;
